Added Purchase, Login, Register and Security

+++ src/main/java/com/bookstore/Service/BookService.java
@@ -5,76 +5,81 @@ import com.bookstore.Entity.BookgenreEntity;
 import com.bookstore.Entity.GenreEntity;
 import com.bookstore.Enum.CoverType;
 import com.bookstore.POJOs.BookPOJO;
+import com.bookstore.POJOs.GenrePOJO;
 import com.bookstore.Repository.BookGenreRepository;
 import com.bookstore.Repository.BookRepository;
 import com.bookstore.Repository.GenreRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import java.net.http.HttpResponse;
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
 public class BookService {
+
     private final BookRepository bookRepository;
     private final GenreService genreService;
     private final BookGenreRepository bookGenreRepository;
 
-    @Autowired
-    public BookService(BookRepository bookRepository, GenreService genreService, GenreRepository genreRepository, BookGenreRepository bookGenreRepository){
+    public BookService(
+            BookRepository bookRepository,
+            GenreService genreService,
+            GenreRepository genreRepository,
+            BookGenreRepository bookGenreRepository
+    ) {
         this.bookRepository = bookRepository;
         this.genreService = genreService;
         this.bookGenreRepository = bookGenreRepository;
     }
 
-    public List<BookPOJO> getAllBooks(){
-        List<BookEntity> bookEntities = bookRepository.findAll();
-        List<BookPOJO> books = new ArrayList<>();
-        for (BookEntity book:bookEntities){
-            List<BookgenreEntity> bookgenreEntities=genreService.getOneBooksGenres(book);
-            List<GenreEntity> genres=new ArrayList<>();
-            for (BookgenreEntity bookgenre:bookgenreEntities){
-                genres.add(bookgenre.getGenre());
-            }
-            books.add(new BookPOJO(book.getIsbn(),book.getTitle(),book.getPrice(),book.getEdition(),book.getCover().toString(),book.getPageNum(),book.getPublisher(),book.getPhysicalSize(),book.getAuthors(),book.getStock(),genres));
-        }
-
-        return books;
+    public List<BookPOJO> getAllBooks() {
+        return bookRepository.findAll()
+                .stream()
+                .map(this::mapBookToPOJO)
+                .toList();
     }
 
     public BookPOJO getBookByISBN(String isbn) {
         BookEntity book = bookRepository.findBookEntityByIsbn(isbn);
-        List<BookgenreEntity> bookgenreEntities=genreService.getOneBooksGenres(book);
-        List<GenreEntity> genres=new ArrayList<>();
-        for (BookgenreEntity bookgenre:bookgenreEntities){
-            genres.add(bookgenre.getGenre());
-        }
-        return new BookPOJO(book.getIsbn(),book.getTitle(),book.getPrice(),book.getEdition(),book.getCover().toString(),book.getPageNum(),book.getPublisher(),book.getPhysicalSize(),book.getAuthors(),book.getStock(),genres);
-
+        return mapBookToPOJO(book);
     }
 
-    public ResponseEntity<String> addBook(BookPOJO book){
-        try{
-            BookEntity bookEntity = new BookEntity(book.getIsbn(),book.getTitle(),book.getPrice(),book.getEdition(), CoverType.valueOf(book.getCover()),book.getPageNum(),book.getPublisher(),book.getPhysicalSize(),book.getAuthors(),book.getStock());
+    public ResponseEntity<String> addBook(BookPOJO book) {
+        try {
+            BookEntity bookEntity = new BookEntity(
+                    book.getIsbn(),
+                    book.getTitle(),
+                    book.getPrice(),
+                    book.getEdition(),
+                    CoverType.valueOf(book.getCover()),
+                    book.getPageNum(),
+                    book.getPublisher(),
+                    book.getPhysicalSize(),
+                    book.getAuthors(),
+                    book.getStock()
+            );
 
-        bookRepository.save(bookEntity);
+            bookRepository.save(bookEntity);
 
-        for(var g: book.getGenres()){
+            for (GenreEntity genre : genreService.getManagedGenresFromPOJOs(book.getGenres())) {
+                bookGenreRepository.save(new BookgenreEntity(bookEntity, genre));
+            }
 
-            bookGenreRepository.save(new BookgenreEntity(bookEntity,g));
-        }
-        return new ResponseEntity<>("Adding new book was successful", HttpStatus.OK);
+            return new ResponseEntity<>("Adding new book was successful", HttpStatus.OK);
         } catch (Exception e) {
-            return new ResponseEntity<>("Adding new book failed: "+e, HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>("Adding new book failed: " + e.getMessage(), HttpStatus.BAD_REQUEST);
         }
     }
 
     public ResponseEntity<String> updateBook(String isbn, BookPOJO updatedEntity) {
-        try{
+        try {
             BookEntity book = bookRepository.findBookEntityByIsbn(isbn);
+
+            if (book == null) {
+                return new ResponseEntity<>("Book not found with ISBN: " + isbn, HttpStatus.NOT_FOUND);
+            }
+
             book.setTitle(updatedEntity.getTitle());
             book.setPrice(updatedEntity.getPrice());
             book.setEdition(updatedEntity.getEdition());
@@ -84,6 +89,7 @@ public class BookService {
             book.setPhysicalSize(updatedEntity.getPhysicalSize());
             book.setAuthors(updatedEntity.getAuthors());
             book.setStock(updatedEntity.getStock());
+
             bookRepository.save(book);
 
             if (updatedEntity.getGenres() != null) {
@@ -92,20 +98,51 @@ public class BookService {
 
             return new ResponseEntity<>("Updating book was successful", HttpStatus.OK);
         } catch (Exception e) {
-            return new ResponseEntity<>("Updating book failed: " + e, HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>("Updating book failed: " + e.getMessage(), HttpStatus.BAD_REQUEST);
         }
     }
 
     public ResponseEntity<String> deleteBook(String isbn) {
-        try{
+        try {
             BookEntity book = bookRepository.findBookEntityByIsbn(isbn);
-            bookRepository.delete(book);
+
+            if (book == null) {
+                return new ResponseEntity<>("Book not found with ISBN: " + isbn, HttpStatus.NOT_FOUND);
+            }
 
             bookGenreRepository.deleteAll(bookGenreRepository.findByBookIsbn(book));
+            bookRepository.delete(book);
 
             return new ResponseEntity<>("Deleting book was successful", HttpStatus.OK);
         } catch (Exception e) {
-            return new ResponseEntity<>("Deleting book failed: " + e, HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>("Deleting book failed: " + e.getMessage(), HttpStatus.BAD_REQUEST);
         }
+    }
+
+    private BookPOJO mapBookToPOJO(BookEntity book) {
+        if (book == null) {
+            return null;
+        }
+
+        List<GenrePOJO> genres = genreService.mapGenresToPOJOs(
+                genreService.getOneBooksGenres(book)
+                        .stream()
+                        .map(BookgenreEntity::getGenre)
+                        .toList()
+        );
+
+        return new BookPOJO(
+                book.getIsbn(),
+                book.getTitle(),
+                book.getPrice(),
+                book.getEdition(),
+                book.getCover() == null ? null : book.getCover().name(),
+                book.getPageNum(),
+                book.getPublisher(),
+                book.getPhysicalSize(),
+                book.getAuthors(),
+                book.getStock(),
+                genres
+        );
     }
 }
